fix: resolve 11 critical frontend/backend integration issues
Frontend fixes:
- Add baseURL and auth interceptors to PQR service
- Configure analytics service with correct env variable
- Implement ISO date conversion helper for analytics
- Fix SelectItem empty string values to 'all'
- Remove infinite loop in AsesoresFilters useEffect
- Add baseURL and auth interceptors to asesores service

Backend fixes:
- Fix UUID to string conversion in auth endpoints
- Simplify Tortoise ORM queries (remove .values(), add .all())
- Fix PQR service cliente.telefono -> cliente.usuario.telefono
- Reorder asesores routes (specific paths before dynamic /{id})
- Replace Tortoise ORM aggregate() with manual calculations
- Change PQRMetrics schema from Decimal to float for proper JSON serialization

All modules (Dashboard, Asesores, PQR, Configuración) now working without errors.

Resolves: #frontend-500-errors

+++ services/core-api/schemas/pqr.py
@@ -70,10 +70,10 @@ class PQRMetrics(BaseModel):
     total_abiertas: int
     total_en_proceso: int
     total_cerradas: int
-    tiempo_promedio_resolucion_horas: Decimal
+    tiempo_promedio_resolucion_horas: float
     pqrs_alta_prioridad: int
     pqrs_criticas: int
-    tasa_resolucion_24h: Decimal  # Porcentaje de PQRs resueltas en menos de 24h
+    tasa_resolucion_24h: float  # Porcentaje de PQRs resueltas en menos de 24h
     distribucion_por_tipo: dict  # {"PETICION": 10, "QUEJA": 5, "RECLAMO": 3}
     distribucion_por_prioridad: dict  # {"BAJA": 8, "MEDIA": 7, "ALTA": 2, "CRITICA": 1}
